+++ lib/file/dicom/image.cpp
@@ -197,9 +197,9 @@ namespace MR {
 
       std::ostream& operator<< (std::ostream& stream, const Image& item)
       {
-        stream << "            " << ( item.instance == UINT_MAX ? 0 : item.instance ) << "#" 
-          << ( item.acq == UINT_MAX ? 0 : item.acq) << ":"
-          << ( item.sequence == UINT_MAX ? 0 : item.sequence ) << " (" 
+        stream << "            " << ( item.instance == SIZE_MAX ? 0 : item.instance ) << "#" 
+          << ( item.acq == SIZE_MAX ? 0 : item.acq) << ":"
+          << ( item.sequence == SIZE_MAX ? 0 : item.sequence ) << " (" 
           << ( item.sequence_name.size() ? item.sequence_name : "?" ) << "), "
           << item.dim[0] << "x" << item.dim[1] << ", "
           << item.pixel_size[0] << "x" << item.pixel_size[1] << " x "
