+++ lib/image/name_parser.cpp
@@ -40,7 +40,7 @@ namespace MR {
       inline bool in_seq (const std::vector<int>& seq, int val) 
       {
         if (seq.size() == 0) return (true);
-        for (uint i = 0; i < seq.size(); i++) 
+        for (size_t i = 0; i < seq.size(); i++) 
           if (seq[i] == val) return (true);
         return (false);
       }
@@ -53,7 +53,7 @@ namespace MR {
 
 
 
-    void NameParser::parse (const std::string& imagename, uint max_num_sequences)
+    void NameParser::parse (const std::string& imagename, size_t max_num_sequences)
     {
       specification = imagename;
       if (Path::is_dir (imagename)) {
@@ -68,7 +68,7 @@ namespace MR {
       try {
         std::string::size_type pos;
         std::string basename = Path::basename (specification);
-        uint num = 0;
+        size_t num = 0;
 
         while ((pos = basename.find_last_of (']')) < std::string::npos && num < max_num_sequences) {
           insert_str (basename.substr (pos+1));
@@ -84,11 +84,11 @@ namespace MR {
         insert_str (basename);
 
 
-        for (uint i = 0; i < array.size(); i++) 
+        for (size_t i = 0; i < array.size(); i++) 
           if (array[i].is_sequence()) 
             if (array[i].sequence().size()) 
-              for (uint n = 0; n < array[i].sequence().size()-1; n++) 
-                for (uint m = n+1; m < array[i].sequence().size(); m++) 
+              for (size_t n = 0; n < array[i].sequence().size()-1; n++) 
+                for (size_t m = n+1; m < array[i].sequence().size(); m++) 
                   if (array[i].sequence()[n] == array[i].sequence()[m]) 
                     throw Exception ("malformed image sequence specifier for image \"" + specification + "\" (duplicate indices)");
       }
@@ -120,7 +120,7 @@ namespace MR {
     std::ostream& operator<< (std::ostream& stream, const NameParser& parser)
     {
       stream << "Image::NameParser: " << parser.specification << "\n";
-      for (uint i = 0; i < parser.array.size(); i++) 
+      for (size_t i = 0; i < parser.array.size(); i++) 
         stream << "  " << i << ": " << parser.array[i] << "\n";
       return (stream);
     }
@@ -137,10 +137,10 @@ namespace MR {
     bool NameParser::match (const std::string& file_name, std::vector<int>& indices) const
     {
       int current = 0;
-      uint num = 0;
+      size_t num = 0;
       indices.resize (seq_index.size());
 
-      for (uint i = 0; i < array.size(); i++) {
+      for (size_t i = 0; i < array.size(); i++) {
         if (array[i].is_string()) {
           if (file_name.substr(current, array[i].string().size()) != array[i].string()) return (false);
           current += array[i].string().size();
@@ -163,21 +163,21 @@ namespace MR {
     void NameParser::calculate_padding (const std::vector<int>& maxvals)
     {
       assert (maxvals.size() == seq_index.size());
-      for (uint n = 0; n < seq_index.size(); n++) 
+      for (size_t n = 0; n < seq_index.size(); n++) 
         assert (maxvals[n] > 0);
 
-      for (uint n = 0; n < seq_index.size(); n++) {
-        uint m = seq_index.size() - 1 - n;
+      for (size_t n = 0; n < seq_index.size(); n++) {
+        size_t m = seq_index.size() - 1 - n;
         NameParserItem& item (array[seq_index[n]]);
         if (item.sequence().size()) {
           if (maxvals[m]) 
-            if (item.sequence().size() != (uint) maxvals[m]) 
+            if (item.sequence().size() != (size_t) maxvals[m]) 
               throw Exception ("dimensions requested in image specifier \"" + specification 
                   + "\" do not match supplied header information");
         }
         else {
           item.sequence().resize (maxvals[m]);
-          for (uint i = 0; i < item.sequence().size(); i++)
+          for (size_t i = 0; i < item.sequence().size(); i++)
             item.sequence()[i] = i;
         }
 
@@ -189,15 +189,15 @@ namespace MR {
 
 
 
-    void NameParserItem::calc_padding (uint maxval)
+    void NameParserItem::calc_padding (size_t maxval)
     {
-      for (uint i = 0; i < sequence().size(); i++) {
+      for (size_t i = 0; i < sequence().size(); i++) {
         assert (sequence()[i] >= 0);
-        if (maxval < (uint) sequence()[i]) maxval = sequence()[i];
+        if (maxval < (size_t) sequence()[i]) maxval = sequence()[i];
       }
 
       seq_length = 1;
-      for (uint num = 10; maxval >= num; num *= 10) 
+      for (size_t num = 10; maxval >= num; num *= 10) 
         seq_length += 1; 
     }
 
@@ -213,8 +213,8 @@ namespace MR {
       assert (indices.size() == seq_index.size());
 
       std::string str;
-      uint n = seq_index.size()-1;
-      for (uint i = 0; i < array.size(); i++) {
+      size_t n = seq_index.size()-1;
+      for (size_t i = 0; i < array.size(); i++) {
         if (array[i].is_string()) str += array[i].string();
         else { 
           str += printf ("%*.*d", array[i].size(), array[i].size(), array[i].sequence()[indices[n]]);
@@ -245,9 +245,9 @@ namespace MR {
       while ((fname = folder->read_name()).size()) {
         if (match (fname, indices)) {
           if (return_seq_index) {
-            for (uint i = 0; i < ndim(); i++) {
+            for (size_t i = 0; i < ndim(); i++) {
               if (sequence(i).size()) {
-                uint n = 0;
+                size_t n = 0;
                 while (indices[i] != sequence(i)[n]) n++; 
                 indices[i] = n;
               }
@@ -263,7 +263,7 @@ namespace MR {
 
     bool ParsedName::operator< (const ParsedName& pn) const
     {
-      for (uint i = 0; i < ndim(); i++) 
+      for (size_t i = 0; i < ndim(); i++) 
         if (index(i) != pn.index(i)) 
           return (index(i) < pn.index(i));
       return (false);
@@ -272,7 +272,7 @@ namespace MR {
 
 
 
-    std::vector<int> ParsedNameList::parse_scan_check (const std::string& specifier, uint max_num_sequences)
+    std::vector<int> ParsedNameList::parse_scan_check (const std::string& specifier, size_t max_num_sequences)
     {
       NameParser parser;
       parser.parse (specifier);
@@ -281,7 +281,7 @@ namespace MR {
       std::sort (begin(), end());
       std::vector<int> dim = count();
 
-      for (uint n = 0; n < dim.size(); n++) 
+      for (size_t n = 0; n < dim.size(); n++) 
         if (parser.sequence(n).size()) 
           if (dim[n] != (int) parser.sequence(n).size()) 
             throw Exception ("number of files found does not match specification \"" + specifier + "\"");
@@ -324,7 +324,7 @@ namespace MR {
       }
 
       std::vector<int> dim ((*this)[0]->ndim(), 0);
-      uint current_entry = 0;
+      size_t current_entry = 0;
       
       count_dim (dim, current_entry, 0);
 
@@ -334,14 +334,14 @@ namespace MR {
 
 
 
-    void ParsedNameList::count_dim (std::vector<int>& dim, uint& current_entry, uint current_dim) const
+    void ParsedNameList::count_dim (std::vector<int>& dim, size_t& current_entry, size_t current_dim) const
     {
       int n;
       bool stop = false;
       RefPtr<const ParsedName> first_entry ((*this)[current_entry]);
 
       for (n = 0; current_entry < size(); n++) {
-        for (uint d = 0; d < current_dim; d++)
+        for (size_t d = 0; d < current_dim; d++)
           if ((*this)[current_entry]->index(d) != first_entry->index(d)) stop = true;
         if (stop) break;
 
@@ -370,7 +370,7 @@ namespace MR {
     std::ostream& operator<< (std::ostream& stream, const ParsedName& pin)
     {
       stream << "[ ";
-      for (uint n = 0; n < pin.ndim(); n++) 
+      for (size_t n = 0; n < pin.ndim(); n++) 
         stream << pin.index(n) << " ";
       stream << "] " << pin.name();
       return (stream);
