+++ lib/image/name_parser.h
@@ -42,16 +42,16 @@ namespace MR {
         std::vector<int>&       sequence () { return (seq); }
         bool                    is_string () const { return (seq_length == 0); }
         bool                    is_sequence () const { return (seq_length != 0); }
-        uint                    size () const { return (seq_length ? seq_length : str.size()); }
+        size_t                  size () const { return (seq_length ? seq_length : str.size()); }
 
-        void                    calc_padding (uint maxval = 0);
+        void                    calc_padding (size_t maxval = 0);
 
         friend std::ostream& operator<< (std::ostream& stream, const NameParserItem& item);
 
       protected:
-        uint                   seq_length;
-        std::string            str;
-        std::vector<int>       seq;
+        size_t           seq_length;
+        std::string      str;
+        std::vector<int> seq;
     };
 
 
@@ -62,35 +62,35 @@ namespace MR {
     class NameParser {
       public:
         NameParser () : folder (NULL) { } 
-        void                       parse (const std::string& imagename, uint max_num_sequences = UINT_MAX);
-        uint                       num () const { return (array.size()); }
+        void                       parse (const std::string& imagename, size_t max_num_sequences = SIZE_MAX);
+        size_t                     num () const { return (array.size()); }
         std::string                spec () const { return (specification); }
-        const NameParserItem&      operator[] (uint i) const { return (array[i]); }
-        const std::vector<int>&    sequence (uint index) const { return (array[seq_index[index]].sequence()); }
-        uint                       ndim () const { return (seq_index.size()); }
-        uint                       index_of_sequence (uint number = 0) const { return (seq_index[number]); }
+        const NameParserItem&      operator[] (size_t i) const { return (array[i]); }
+        const std::vector<int>&    sequence (size_t index) const { return (array[seq_index[index]].sequence()); }
+        size_t                     ndim () const { return (seq_index.size()); }
+        size_t                     index_of_sequence (size_t number = 0) const { return (seq_index[number]); }
 
-        bool                       match (const std::string& file_name, std::vector<int>& indices) const;
-        void                       calculate_padding (const std::vector<int>& maxvals);
-        std::string                name (const std::vector<int>& indices);
-        std::string                get_next_match (std::vector<int>& indices, bool return_seq_index = false);
+        bool        match (const std::string& file_name, std::vector<int>& indices) const;
+        void        calculate_padding (const std::vector<int>& maxvals);
+        std::string name (const std::vector<int>& indices);
+        std::string get_next_match (std::vector<int>& indices, bool return_seq_index = false);
 
         friend std::ostream& operator<< (std::ostream& stream, const NameParser& parser);
 
       private:
         std::vector<NameParserItem> array;
-        std::vector<uint>           seq_index;
+        std::vector<size_t>         seq_index;
         std::string                 folder_name;
         std::string                 specification;
         std::string                 current_name;
         Path::Dir*                  folder;
 
-        void                        insert_str (const std::string& str) {
+        void insert_str (const std::string& str) {
           NameParserItem item;
           item.set_str (str);
           array.insert (array.begin(), item);
         }
-        void                        insert_seq (const std::string& str) {
+        void insert_seq (const std::string& str) {
           NameParserItem item;
           item.set_seq (str);
           array.insert (array.begin(), item);
@@ -113,9 +113,9 @@ namespace MR {
       public:
         ParsedName (const std::string& name, const std::vector<int>& index) : indices (index), filename (name) { }
 
-        std::string         name () const { return (filename); }
-        uint                ndim () const { return (indices.size()); }
-        int                 index (uint num) const { return (indices[num]); }
+        std::string name () const { return (filename); }
+        size_t      ndim () const { return (indices.size()); }
+        int         index (size_t num) const { return (indices[num]); }
 
         bool                 operator< (const ParsedName& pn) const;
         friend std::ostream& operator<< (std::ostream& stream, const ParsedName& pin);
@@ -127,15 +127,15 @@ namespace MR {
 
     class ParsedNameList : public std::vector< RefPtr<ParsedName> > {
       protected:
-        void               count_dim (std::vector<int>& dim, uint& current_entry, uint current_dim) const;
-        uint               max_name_size;
+        void   count_dim (std::vector<int>& dim, size_t& current_entry, size_t current_dim) const;
+        size_t max_name_size;
 
       public:
-        std::vector<int>    parse_scan_check (const std::string& specifier, uint max_num_sequences = UINT_MAX);
-        void                scan (NameParser& parser);
+        std::vector<int> parse_scan_check (const std::string& specifier, size_t max_num_sequences = SIZE_MAX);
+        void             scan (NameParser& parser);
 
-        std::vector<int>    count () const;
-        uint                biggest_filename_size () const { return (max_name_size); }
+        std::vector<int> count () const;
+        size_t           biggest_filename_size () const { return (max_name_size); }
     };
