+++ lib/mrtrix.cpp
@@ -23,7 +23,6 @@
 #include <gsl/gsl_version.h>
 
 #include "mrtrix.h"
-#include "image/object.h"
 
 namespace MR {
 
@@ -31,9 +30,9 @@ namespace MR {
    *                     MRtrix version information                      *
    ************************************************************************/
 
-  const uint mrtrix_major_version = MRTRIX_MAJOR_VERSION;
-  const uint mrtrix_minor_version = MRTRIX_MINOR_VERSION;
-  const uint mrtrix_micro_version = MRTRIX_MICRO_VERSION;
+  const size_t mrtrix_major_version = MRTRIX_MAJOR_VERSION;
+  const size_t mrtrix_minor_version = MRTRIX_MINOR_VERSION;
+  const size_t mrtrix_micro_version = MRTRIX_MICRO_VERSION;
 
 
   /************************************************************************
